Make a separate file for AnimatedWorm
